ACCUMULO-1537 fixed NPE in MAC
git-svn-id: https://svn.apache.org/repos/asf/accumulo/trunk@1502381 13f79535-47bb-0310-9956-ffa450edef68

+++ minicluster/src/main/java/org/apache/accumulo/minicluster/MiniAccumuloCluster.java
@@ -244,11 +244,14 @@ public class MiniAccumuloCluster {
 
     File nativeMap = new File(config.getLibDir().getAbsolutePath() + "/native/map");
     nativeMap.mkdirs();
-    String testRoot = new File(new File(System.getProperty("user.dir")).getParent() + "/server/src/main/c++/nativeMap").getAbsolutePath();
-    for (String file : new File(testRoot).list()) {
-      File src = new File(testRoot, file);
-      if (src.isFile() && file.startsWith("libNativeMap"))
-        FileUtils.copyFile(src, new File(nativeMap, file));
+    File testRoot = new File(new File(new File(System.getProperty("user.dir")).getParent() + "/server/src/main/c++/nativeMap").getAbsolutePath());
+    
+    if (testRoot.exists()) {
+      for (String file : testRoot.list()) {
+        File src = new File(testRoot, file);
+        if (src.isFile() && file.startsWith("libNativeMap"))
+          FileUtils.copyFile(src, new File(nativeMap, file));
+      }
     }
   }
